+++ cli/Valet/Brew.php
@@ -11,6 +11,7 @@ class Brew
     // This is the array of PHP versions that Valet will attempt to install/configure when requested
     const SUPPORTED_PHP_VERSIONS = [
         'php',
+        'php@8.3',
         'php@8.2',
         'php@8.1',
         'php@8.0',
