+++ cli/Valet/Drivers/ValetDriver.php
@@ -45,8 +45,9 @@ abstract class ValetDriver
         $drivers = array_merge($drivers, static::customDrivers());
 
         // Queue Valet-shipped drivers
+        $drivers[] = 'Specific\StatamicValetDriver';
         $drivers[] = 'LaravelValetDriver';
-        $drivers = array_merge($drivers, $specificDrivers);
+        $drivers = array_unique(array_merge($drivers, $specificDrivers));
         $drivers[] = 'BasicWithPublicValetDriver';
         $drivers[] = 'BasicValetDriver';
