+++ composer.json
@@ -34,11 +34,14 @@
             "tests/Drivers/BaseDriverTestCase.php"
         ]
     },
+    "conflict": {
+        "mnapoli/silly": ">=1.8.1 <1.8.3"
+    },
     "require": {
         "php": "^7.1|^8.0",
         "illuminate/collections": "^8.0|^9.0|^10.0",
         "illuminate/container": "~5.1|^6.0|^7.0|^8.0|^9.0|^10.0",
-        "mnapoli/silly": "^1.0",
+        "mnapoli/silly": "^1.5",
         "symfony/console": "^3.0|^4.0|^5.0|^6.0",
         "symfony/process": "^3.0|^4.0|^5.0|^6.0",
         "guzzlehttp/guzzle": "^6.0|^7.4",
